fixing circular deps with asset-swapper assert
