+++ src/asset-swapper/utils/utils.ts
@@ -1,7 +1,8 @@
 import { CommonOrderFields, FillQuoteTransformerOrderType, LimitOrderFields } from '@0x/protocol-utils';
 import { BigNumber } from '@0x/utils';
+import { assert as sharedAssert } from '@0x/assert';
 
-import { NativeOrderFillableAmountFields, SignedNativeOrder } from '../types';
+import { NativeOrderFillableAmountFields, SignedNativeOrder, Orderbook } from '../types';
 
 import { ZERO_AMOUNT } from './market_operation_utils/constants';
 
@@ -86,3 +87,21 @@ export function getNativeAdjustedFillableAmountsFromMakerAmount(
                 : ZERO_AMOUNT,
     };
 }
+
+export const assert = {
+    ...sharedAssert,
+    isValidOrderbook(variableName: string, orderFetcher: Orderbook): void {
+        sharedAssert.isFunction(`${variableName}.getOrdersAsync`, orderFetcher.getOrdersAsync.bind(orderFetcher));
+        sharedAssert.isFunction(
+            `${variableName}.getBatchOrdersAsync`,
+            orderFetcher.getBatchOrdersAsync.bind(orderFetcher),
+        );
+    },
+    isValidPercentage(variableName: string, percentage: number): void {
+        assert.isNumber(variableName, percentage);
+        assert.assert(
+            percentage >= 0 && percentage <= 1,
+            `Expected ${variableName} to be between 0 and 1, but is ${percentage}`,
+        );
+    },
+};
